+++ data-raw/creating_reuters_txt.R
@@ -1,4 +1,5 @@
 library(tm)
+
 reuters_files <- system.file("texts", "crude", package = "tm")
 reuters_vcorpus <- VCorpus(DirSource(reuters_files), list(reader = readReut21578XMLasPlain))
 reuters_txt <- unname(unlist(lapply(reuters_vcorpus, function(x) gsub("\\s+", " ", gsub("\\n", " ", x[["content"]])))))